+++ indy-repository-manager/src/main/java/org/jboss/pnc/indyrepositorymanager/RepositoryManagerDriver.java
@@ -270,7 +270,7 @@ public class RepositoryManagerDriver implements RepositoryManager {
     }
 
     @Override
-    public RepositoryManagerResult collectRepoManagerResult(Integer id) throws RepositoryManagerException {
+    public RepositoryManagerResult collectRepoManagerResult(Long id) throws RepositoryManagerException {
         BuildRecord br = buildRecordRepository.findByIdFetchProperties(id);
         if (br == null) {
             return null;
@@ -318,7 +318,7 @@ public class RepositoryManagerDriver implements RepositoryManager {
             Map<String, String> genericParameters) throws IndyClientException {
 
         String buildContentId = execution.getBuildContentId();
-        int id = execution.getId();
+        long id = execution.getId();
 
         // if the build-level group doesn't exist, create it.
         StoreKey groupKey = new StoreKey(packageType, StoreType.group, buildContentId);
@@ -402,7 +402,7 @@ public class RepositoryManagerDriver implements RepositoryManager {
     private void addExtraConstituents(
             String packageType,
             List<ArtifactRepository> repositories,
-            int buildId,
+            long buildId,
             String buildContentId,
             Indy indy,
             Group buildGroup) throws IndyClientException {
