+++ rest-new/src/main/java/org/jboss/pnc/rest/endpoints/BuildEndpointImpl.java
@@ -91,7 +91,7 @@ public class BuildEndpointImpl implements BuildEndpoint {
     @Inject
     private BrewPusher brewPusher;
 
-    private EndpointHelper<Integer, Build, BuildRef> endpointHelper;
+    private EndpointHelper<Long, Build, BuildRef> endpointHelper;
 
     @PostConstruct
     public void init() {
