+++ spi/src/main/java/org/jboss/pnc/spi/coordinator/BuildTask.java
@@ -41,7 +41,7 @@ public class BuildTask {
 
     private static final Logger userLog = LoggerFactory.getLogger("org.jboss.pnc._userlog_.build-task");
 
-    private final Integer id;
+    private final long id;
     private final BuildConfigurationAudited buildConfigurationAudited; // TODO decouple DB entity
 
     @Getter
@@ -88,7 +88,7 @@ public class BuildTask {
             User user,
             Date submitTime,
             BuildSetTask buildSetTask,
-            int id,
+            long id,
             Integer buildConfigSetRecordId,
             ProductMilestone productMilestone,
             String contentId,
@@ -235,7 +235,7 @@ public class BuildTask {
         this.hasFailed = hasFailed;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
@@ -291,7 +291,7 @@ public class BuildTask {
             BuildConfigurationAudited buildConfigurationAudited,
             BuildOptions buildOptions,
             User user,
-            int buildTaskId,
+            long buildTaskId,
             BuildSetTask buildSetTask,
             Date submitTime,
             ProductMilestone productMilestone,
